task-4/server/requirements.txt
autopep8==1.4.2
dj-database-url==0.5.0
Django==2.1.3
django-cors-headers==2.4.0
djangorestframework==3.9.0
pep8==1.7.1
psycopg2==2.7.5
pycodestyle==2.4.0
pytz==2018.7
